fix(prover): resolve import cycles in test files caused by serde registry
Problem:
Multiple test files failed with 'import cycle not allowed in test'.

Root Cause:
New protocol/serde/impl_registry.go imports many packages for type
registration, and protocol/internal/testtools/testcase.go now imports
serde for serialization tests. This creates cycles:
  pkg (test) → testtools → serde → pkg

Fix:
Convert to external test packages (package foo_test), the idiomatic Go
pattern for breaking import cycles in tests.

Affected packages:
- protocol/dedicated
- protocol/compiler/splitextension
- protocol/compiler/mpts"
