+++ prover/protocol/compiler/mpts/multipoint_to_singlepoint_external_test.go
@@ -1,0 +1,121 @@
+package mpts_test
+
+import (
+	"math/rand/v2"
+	"testing"
+
+	"github.com/consensys/linea-monorepo/prover/maths/common/smartvectors"
+	"github.com/consensys/linea-monorepo/prover/maths/field"
+	"github.com/consensys/linea-monorepo/prover/maths/field/fext"
+
+	"github.com/consensys/linea-monorepo/prover/protocol/column/verifiercol"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/dummy"
+	"github.com/consensys/linea-monorepo/prover/protocol/compiler/mpts"
+	"github.com/consensys/linea-monorepo/prover/protocol/ifaces"
+	"github.com/consensys/linea-monorepo/prover/protocol/internal/testtools"
+	"github.com/consensys/linea-monorepo/prover/protocol/wizard"
+	"github.com/consensys/linea-monorepo/prover/utils"
+)
+
+func TestWithProfile(t *testing.T) {
+	var rng = rand.New(utils.NewRandSource(0)) // #nosec G404 -- test only
+
+	testcases := []struct {
+		Profile []int
+		UTC     *testtools.UnivariateTestcase
+	}{
+		{
+			Profile: []int{3},
+			UTC: &testtools.UnivariateTestcase{
+				NameStr: "profile-3",
+				Polys: []smartvectors.SmartVector{
+					smartvectors.ForTest(1, 2, 3, 4),
+				},
+				QueryXs: []fext.Element{
+					fext.PseudoRand(rng),
+				},
+				QueryPols: [][]int{
+					{0},
+				},
+			},
+		},
+	}
+
+	for _, tc := range testcases {
+
+		suite := []func(*wizard.CompiledIOP){
+			mpts.Compile(mpts.WithNumColumnProfileOpt(tc.Profile, 0)),
+			dummy.Compile,
+		}
+
+		t.Run(tc.UTC.NameStr, func(t *testing.T) {
+			testtools.RunTestcase(t, tc.UTC, suite)
+		})
+	}
+
+}
+
+func TestCompiler(t *testing.T) {
+
+	suite := []func(*wizard.CompiledIOP){
+		mpts.Compile(),
+		dummy.Compile,
+	}
+
+	for _, tc := range testtools.ListOfUnivariateTestcasesPositive {
+		t.Run(tc.Name(), func(t *testing.T) {
+			testtools.RunTestcase(t, tc, suite)
+		})
+	}
+}
+
+func TestCompilerWithGnark(t *testing.T) {
+
+	suite := []func(*wizard.CompiledIOP){
+		mpts.Compile(),
+		dummy.Compile,
+	}
+
+	for _, tc := range testtools.ListOfUnivariateTestcasesPositive {
+		t.Run(tc.Name(), func(t *testing.T) {
+			testtools.RunTestShouldPassWithGnarkKoala(t, tc, suite)
+		})
+	}
+}
+
+func TestWithVerifierCol(t *testing.T) {
+
+	suite := []func(*wizard.CompiledIOP){
+		mpts.Compile(),
+		dummy.Compile,
+	}
+
+	testcases := []*testtools.AnonymousTestcase{
+		{
+			NameStr: "with-constant-col",
+			DefineFunc: func(comp *wizard.CompiledIOP) {
+				u := verifiercol.NewConstantCol(field.Zero(), 8, "")
+				comp.InsertUnivariate(0, "U", []ifaces.Column{u})
+			},
+			AssignFunc: func(run *wizard.ProverRuntime) {
+				run.AssignUnivariateExt("U", fext.Zero(), fext.Zero())
+			},
+		},
+		{
+			NameStr: "with-constant-col-2",
+			DefineFunc: func(comp *wizard.CompiledIOP) {
+				u := verifiercol.NewConstantCol(field.NewElement(42), 4, "")
+				comp.InsertUnivariate(0, "U", []ifaces.Column{u})
+			},
+			AssignFunc: func(run *wizard.ProverRuntime) {
+				run.AssignUnivariateExt("U", fext.Zero(), fext.NewFromUint(42, 0, 0, 0))
+			},
+		},
+	}
+
+	for i, tc := range testcases {
+		t.Run(tc.NameStr, func(t *testing.T) {
+			testtools.RunTestcase(t, testcases[i], suite)
+		})
+	}
+}
