+++ prover/protocol/dedicated/repeated_pattern_test.go
@@ -1,4 +1,4 @@
-package dedicated
+package dedicated_test
 
 import (
 	"testing"
@@ -8,18 +8,19 @@ import (
 	"github.com/consensys/linea-monorepo/prover/maths/field"
 	"github.com/consensys/linea-monorepo/prover/protocol/column/verifiercol"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/dummy"
+	"github.com/consensys/linea-monorepo/prover/protocol/dedicated"
 	"github.com/consensys/linea-monorepo/prover/protocol/ifaces"
 	"github.com/consensys/linea-monorepo/prover/protocol/internal/testtools"
 	"github.com/consensys/linea-monorepo/prover/protocol/wizard"
 	"github.com/stretchr/testify/assert"
 )
 
-// RepeatedPatternTestcase represents a test case for [RepeatedPattern].
+// RepeatedPatternTestcase represents a test case for [dedicated.RepeatedPattern].
 type RepeatedPatternTestcase struct {
 	name     string
 	IsActive smartvectors.SmartVector
 	Pattern  []field.Element
-	rp       *RepeatedPattern
+	rp       *dedicated.RepeatedPattern
 }
 
 // ListOfRepeatedPatternTestcase is a list of [RepeatedPatternTestcase].
@@ -53,7 +54,7 @@ var ListOfRepeatedPatternTestcase = []*RepeatedPatternTestcase{
 
 func (rp *RepeatedPatternTestcase) Define(comp *wizard.CompiledIOP) {
 	isActive := comp.InsertCommit(0, ifaces.ColID(rp.name)+"_ACTIVE", rp.IsActive.Len(), true)
-	rp.rp = NewRepeatedPattern(comp, 0, rp.Pattern, isActive, "TEST")
+	rp.rp = dedicated.NewRepeatedPattern(comp, 0, rp.Pattern, isActive, "TEST")
 }
 
 func (rp *RepeatedPatternTestcase) Assign(run *wizard.ProverRuntime) {
@@ -83,10 +84,10 @@ func TestRepeatedPattern(t *testing.T) {
 
 func TestRepeatedPatWithVerifCol(t *testing.T) {
 
-	var rp *RepeatedPattern
+	var rp *dedicated.RepeatedPattern
 	define := func(b *wizard.Builder) {
 		pattern := vector.ForTest(1, 2, 3)
-		rp = NewRepeatedPattern(b.CompiledIOP, 0, pattern, verifiercol.NewConstantCol(field.One(), 32, "active"), "TEST")
+		rp = dedicated.NewRepeatedPattern(b.CompiledIOP, 0, pattern, verifiercol.NewConstantCol(field.One(), 32, "active"), "TEST")
 	}
 	prove := func(run *wizard.ProverRuntime) {
 		rp.Assign(run)
